backend/pyproject.toml
[project]
name = "veritasad-backend"
version = "1.0.0"
description = "Neural network-based advertising detection system API"
authors = [{name = "VeritasAd Team", email = "dabinayo@pm.me"}]
requires-python = ">=3.11"                  # можно поднять до >=3.12 если whisper не нужен
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # FastAPI & web
    "fastapi==0.128.0",                     # самая свежая на декабрь 2025
    "uvicorn[standard]==0.32.1",            # обновлено (0.32.0 → 0.32.1 или новее)
    "pydantic==2.10.0",                     # 2.9.2 → 2.10+
    "pydantic-settings==2.7.0",             # актуально
    "python-dotenv==1.0.1",
    "python-multipart==0.0.20",             # сильно обновилась

    # Async database
    "sqlalchemy[asyncio]==2.0.36",          # свежая 2.0.x
    "asyncpg==0.30.0",                      # обновлено с 0.29 → 0.30
    "alembic==1.14.0",
    "greenlet==3.1.1",                      # обычно сам подтягивается

    # Redis & task queue
    "redis[hiredis]==5.2.0",
    "celery[redis]==5.4.0",                 # стабильная
    "flower==2.0.1",

    # Rate limiting
    "slowapi==0.1.9",
    "limits==3.13.0",

    # Logging
    "structlog==24.4.0",
    "python-json-logger==3.1.0",

    # Video processing
    "yt-dlp==2025.12.8",                    # самая свежая стабильная на конец 2025
    "opencv-python-headless==4.11.0.86",    # headless предпочтительнее для серверов
    "ffmpeg-python==0.2.0",                 # без изменений, альтернативы почти нет

    # Machine Learning
    "torch==2.9.1",                         # последняя стабильная на ноябрь 2025
    "transformers==4.57.3",                 # свежая
    "sentencepiece==0.2.1",
    # "openai-whisper==20231117",           # ← закомментировано! Проблемы на 3.13+
    # Рекомендация: используйте faster-whisper или оставайтесь на Python 3.12
    "faster-whisper==1.0.3",
    "accelerate==1.1.1",

    # Audio processing
    "librosa==0.10.2.post1",                # можно попробовать 0.11.x если нужно

    # PDF generation
    "reportlab==4.2.5",
    "pillow==11.0.0",

    # OCR
    "pytesseract==0.3.13",                  # последняя, runtime может быть нестабилен на 3.14

    # Utilities & security
    "numpy==2.1.2",                         # 1.26 → 2.1+ (очень рекомендуется)
    "requests==2.32.3",
    "httpx==0.28.0",                        # обновлено с 0.27.2
    "aiofiles==24.1.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",                        # обновлено
    "pytest-asyncio==0.25.0",               # свежее
    "pytest-cov==6.0.0",
    "httpx==0.28.0",
    "black==24.10.0",
    "ruff==0.7.4",                          # можно обновить до 0.8+ если стабильно
    "mypy==1.13.0",
    "pre-commit==4.0.1",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/__pycache__/*", "*/venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]